fix: Match encoding/json integer parsing in both FDv2 build variants
readInt went through jreader.Float64, so it truncated fractional versions and
produced an implementation-defined (CPU-architecture-dependent) result for a
value outside the int range -- diverging from the reflection decode the
launchdarkly_easyjson build uses.

Split the scalar event decoders by build tag so both variants agree with
encoding/json: the default build parses the raw number literal with strconv
(rejecting fractional, exponent, and out-of-range values while preserving large
integers exactly and treating null as zero), and the easyjson build decodes each
scalar event with json.Unmarshal into the subsystems types. The payload-
transferred selector is kept lenient in both builds to match Selector's
float64-based decode.

Adds a differential test that checks every integer field against an
encoding/json oracle under both build tags.
